SCALE-SEAM (103) — acceptance gates, client.ts 589 -> 572
permitReadiness, diligenceReadiness, handoverAcceptance and validate ->
acceptanceGates.ts: "will an outside party accept this project, and what is
blocking it".

NEITHER THE ROUTE NOR THE AUDIENCE PRODUCES THIS GROUPING. The four routes are
/permit/readiness, /diligence/readiness, /handover/acceptance and /validate --
four different prefixes, so a prefix grouping takes one each and the set never
forms; the shared leaf word "readiness" reaches only two of the four. The
deciding parties are an AHJ, an investor, an owner and an IDS checker -- four
different audiences, so grouping by reader fails too.

What forms it is the RETURN SHAPE: each collapses the whole project to a single
accept/refuse verdict (verdict / go / accepted / status:"pass"|"fail") and then
enumerates what withholds it. (102) was carried by a prefix that actively
DISAGREED with the seam; this one by a vocabulary that says nothing at all --
four names sharing no words can still be one question, which is the inverse of
the error (85), (89) and annotate.ts each recorded, where a shared mechanism
looked like a shared question.

THE TWO EXCLUSIONS DID THE WORK. spineTraceability is the closest miss -- same
domain, adjacent in the file, equally project-scoped -- and returns
coverage/gaps/chain with NO verdict field at all: it maps completeness for a
human rather than deciding acceptance. editPrecheck DOES return a verdict, but
judges a PENDING ACTION ("may I run this recipe with these params"), is remedied
by changing the params you are about to submit, and sits beside addCurtainWall
as the precheck for editIfc.

ONE FALSE POSITIVE, RECORDED BECAUSE IT WAS INVISIBLE. A scan for verdict-shaped
returns also flagged collabSnapshot -- but the match came from the doc comment
introducing permitReadiness, the NEXT method. A method-body splitter that runs
to the next header swallows the comment belonging to that header, so the
population silently inherits its neighbour's vocabulary; a count of six looked
entirely checked. Only reading each candidate caught it.

Also files handoverAcceptance, which an earlier slice had parked under an
explicit UNFILED note asking for it to be placed by what it ANSWERS rather than
by what it sits next to. That note is narrowed to two entries, not deleted --
the other two are still genuinely unfiled, and a note that silently loses
entries is how earlier slices lost methods.

THE DOC-STRAND GATE CAUGHT A DEFECT IN THIS SLICE'S OWN NEW FILE: a /** */ doc
comment on the Ctor type sat directly below the module header, which that gate
reads as a stranded comment -- the header carrying the whole witness was one
line from reading as documentation of a type alias. It is a // comment now,
which is why counterpartyRisk.ts has none there either. The fast checks (tsc,
eslint) were clean while this was broken; the suite is where it lived.

Verified: 2072 web tests / 205 files, tsc + eslint clean, whole-tree ruff clean,
ratchet mutation-checked at 571 (both the growth and chain-end assertions fire),
48-entry narrative chain unbroken, test_route_reachability + test_reachable +
test_import_cycles + test_doc_substance + test_claude_md_gates + test_ruff_scope
+ test_declared_imports green.
